Move Login files to keyword package
