+++ presentation/src/main/res/navigation/keyword_activity_nav.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.foundy.presentation.view.login.LoginFragment"
+        android:name="com.foundy.presentation.view.keyword.LoginFragment"
         android:label="@string/notification_keyword"
         tools:layout="@layout/fragment_login">
         <action
